Finished the most basic product
I split the main.js into more javascript files so it would be so big.
am using a Single-page-application
have no errors or warnings from ESLint
