--- server/public/footer.html
@@ -1,3 +1,0 @@
-<footer>
-  <p>© 2025 SPA Auth Project</p>
-</footer>
