--- server/public/header.html
@@ -1,10 +1,0 @@
-<header>
-  <div class="logo">SPA Auth</div>
-  <nav>
-    <a href="#/">Home</a>
-    <a href="#/about">About</a>
-    <a href="#/dashboard">Dashboard</a>
-    <a href="#/login">Sign In</a>
-    <a href="#/signup">Sign Up</a>
-  </nav>
-</header>
